+++ library/LibUtilities/Foundations/NodalQuadElec.cpp
@@ -32,12 +32,7 @@
 //
 ///////////////////////////////////////////////////////////////////////////////
 
-#include <LibUtilities/BasicConst/NektarUnivConsts.hpp>
-#include <LibUtilities/BasicUtils/ErrorUtil.hpp>
-#include <LibUtilities/BasicUtils/SharedArray.hpp>
 #include <LibUtilities/Foundations/NodalQuadElec.h>
-#include <LibUtilities/Foundations/NodalUtil.h>
-#include <LibUtilities/Foundations/Points.h>
 
 namespace Nektar
 {
@@ -50,20 +45,20 @@ bool NodalQuadElec::initPointsManager[] = {PointsManager().RegisterCreator(
 void NodalQuadElec::v_CalculatePoints()
 {
     // Allocate the storage for points
-    unsigned int numPoints = GetNumPoints();
+    size_t numPoints = GetNumPoints();
 
     PointsKey e(numPoints, eGaussLobattoLegendre);
     PointsManager()[e]->GetPoints(m_e0);
     m_ew = PointsManager()[e]->GetW();
 
-    for (int i = 0; i < 2; i++)
+    for (size_t i = 0; i < 2; i++)
     {
         m_points[i] = Array<OneD, DataType>(numPoints * numPoints);
     }
 
-    for (int j = 0, ct = 0; j < numPoints; j++)
+    for (size_t j = 0, ct = 0; j < numPoints; j++)
     {
-        for (int i = 0; i < numPoints; i++, ct++)
+        for (size_t i = 0; i < numPoints; i++, ct++)
         {
             m_points[0][ct] = m_e0[i];
             m_points[1][ct] = m_e0[j];
@@ -73,13 +68,13 @@ void NodalQuadElec::v_CalculatePoints()
 
 void NodalQuadElec::v_CalculateWeights()
 {
-    unsigned int numPoints = GetNumPoints();
+    size_t numPoints = GetNumPoints();
 
     m_weights = Array<OneD, DataType>(numPoints * numPoints);
 
-    for (int j = 0, ct = 0; j < numPoints; j++)
+    for (size_t j = 0, ct = 0; j < numPoints; j++)
     {
-        for (int i = 0; i < numPoints; i++, ct++)
+        for (size_t i = 0; i < numPoints; i++, ct++)
         {
             m_weights[ct] = m_ew[i] * m_ew[j];
         }
